Fixed a bug in decimal to segment encoder

+++ ava/hardware.segmentdisplay.h
@@ -23,7 +23,7 @@ private:
     static const uint8_t _hexToSegmentMap[];
 
     uint8_t getHexDigitSegmentMap(uint8_t n);
-    uint8_t getDecDigitSegmentMap(uint8_t n);
+    uint8_t getDecDigitSegmentMap(uint16_t n);
 
 };
